wip 3 of globus auth
