+++ internal/webserver/auth.go
@@ -8,6 +8,7 @@ import (
 	"time"
 
 	"github.com/SwissOpenEM/Ingestor/internal/task"
+	"github.com/SwissOpenEM/Ingestor/internal/webserver/globusauth"
 	"github.com/SwissOpenEM/Ingestor/internal/webserver/randomfuncs"
 	"github.com/coreos/go-oidc/v3/oidc"
 	"github.com/gin-contrib/sessions"
@@ -24,22 +25,6 @@ func (i *IngestorWebServerImplemenation) GetLogin(ctx context.Context, request G
 		return GetLogin302Response{}, errors.New("CANT CONVERT")
 	}
 	authSession := sessions.DefaultMany(ginCtx, "auth")
-	userSession := sessions.DefaultMany(ginCtx, "user")
-
-	// check if already logged-in
-	if val, ok := userSession.Get("expires_at").(string); ok {
-		expiry, _ := time.Parse(time.RFC3339Nano, val)
-		if time.Now().Before(expiry) {
-			if !loggedIntoGlobus(ginCtx) {
-
-			}
-			return GetLogin302Response{
-				Headers: GetLogin302ResponseHeaders{
-					Location: i.frontend.origin + i.frontend.redirectPath,
-				},
-			}, nil
-		}
-	}
 
 	// generate state, verifier and nonce
 	state, err := randomfuncs.GenerateRandomString(16)
@@ -180,12 +165,16 @@ func (i *IngestorWebServerImplemenation) GetCallback(ctx context.Context, reques
 		return GetCallback500TextResponse(fmt.Sprintf("can't set user session: %s", err.Error())), nil
 	}
 
-	// globus login (if using globus)
+	// globus redirect for logging-in (if using globus)
 	if i.taskQueue.GetTransferMethod() == task.TransferGlobus {
+		// revoke session with globus, if we have one ongoing
+		if globusauth.TestGlobusCookie(ginCtx) {
+			globusauth.Logout(ginCtx, *i.globusAuthConf)
+		}
 		return globusCallbackRedirect(ctx, i.globusAuthConf)
 	}
 
-	// reply
+	// standard redirect to frontend if there's nothing else to do
 	return GetCallback302Response{
 		Headers: GetCallback302ResponseHeaders{
 			Location: i.frontend.origin + i.frontend.redirectPath,
@@ -212,7 +201,7 @@ func (i *IngestorWebServerImplemenation) GetLogout(ctx context.Context, request 
 	}
 
 	if i.taskQueue.GetTransferMethod() == task.TransferGlobus {
-		err = globusLogout(ginCtx, *i.globusAuthConf)
+		err = globusauth.Logout(ginCtx, *i.globusAuthConf)
 		if err != nil {
 			return GetLogout500TextResponse(err.Error()), nil
 		}
@@ -282,25 +271,66 @@ func (i *IngestorWebServerImplemenation) GetUserinfo(ctx context.Context, reques
 	}, nil
 }
 
-/*func getGlobusClientFromSession(ctx *gin.Context, conf *oauth2.Config) (globus.GlobusClient, error) {
-	globusSession := sessions.DefaultMany(ctx, "globus")
-	refreshToken, ok := globusSession.Get("refresh_token").(string)
+// this is the callback endpoint for handling the globus code exchange
+func (i *IngestorWebServerImplemenation) GetGlobusCallback(ctx context.Context, request GetGlobusCallbackRequestObject) (GetGlobusCallbackResponseObject, error) {
+	ginCtx, ok := ctx.(*gin.Context)
 	if !ok {
-		return globus.GlobusClient{}, fmt.Errorf("globus session has expired")
+		return GetGlobusCallback500TextResponse("can't access context"), nil
+	}
+	authSession := sessions.DefaultMany(ginCtx, "auth")
+
+	state, ok1 := authSession.Get("state").(string)
+	verifier, ok2 := authSession.Get("verifier").(string)
+	if !ok1 || !ok2 {
+		return GetGlobusCallback400TextResponse("auth session has expired or is invalid"), nil
 	}
 
-	newToken, err := conf.TokenSource(ctx, &oauth2.Token{RefreshToken: refreshToken}).Token()
+	// delete auth session
+	authSession.Delete("state")
+	authSession.Delete("verifier")
+	authSession.Options(sessions.Options{
+		HttpOnly: true,
+		Secure:   ginCtx.Request.TLS != nil,
+		MaxAge:   -1,
+	})
+	err := authSession.Save()
 	if err != nil {
-		return globus.GlobusClient{}, fmt.Errorf("can't refresh token: %s", err.Error())
+		return GetGlobusCallback500TextResponse(err.Error()), nil
 	}
 
-	globusSession.Set("refresh_token", newToken.RefreshToken)
-	globusSession.Save()
+	if request.Params.State != state {
+		return GetGlobusCallback400TextResponse("invalid state"), nil
+	}
 
-	return globus.HttpClientToGlobusClient(conf.Client(ctx, &oauth2.Token{
-		TokenType:   newToken.TokenType,
-		AccessToken: newToken.AccessToken,
-		Expiry:      newToken.Expiry,
-		ExpiresIn:   newToken.ExpiresIn,
-	})), nil
-}*/
+	// exchange authorization code for accessToken
+	oauthToken, err := i.globusAuthConf.Exchange(
+		ctx,
+		request.Params.Code,
+		oauth2.AccessTypeOffline,
+		oauth2.VerifierOption(verifier),
+	)
+	if err != nil {
+		return GetGlobusCallback400TextResponse(fmt.Sprintf("code exchange failed: %s", err.Error())), nil
+	}
+
+	globusauth.SetTokenCookie(ginCtx, oauthToken.RefreshToken, oauthToken.AccessToken, oauthToken.Expiry, i.sessionDuration)
+
+	return GetGlobusCallback302Response{
+		Headers: GetGlobusCallback302ResponseHeaders{
+			Location: i.frontend.origin + i.frontend.redirectPath,
+		},
+	}, nil
+}
+
+func globusCallbackRedirect(ctx context.Context, globusAuthConf *oauth2.Config) (GetCallbackResponseObject, error) {
+	redirectUrl, err := globusauth.GetRedirectUrl(ctx, globusAuthConf)
+	if err != nil {
+		return GetCallback500TextResponse(err.Error()), nil
+	}
+
+	return GetCallback302Response{
+		Headers: GetCallback302ResponseHeaders{
+			Location: redirectUrl,
+		},
+	}, nil
+}
